remove oldConfig for newConfig
